+++ docker_env
@@ -1,7 +1,7 @@
 DOCKER_REGISTRY=ghcr.io
 DOCKER_ORG=stactools-packages
-DOCKER_REPO=ephemeral
+DOCKER_REPO=nrcan-radarsat1
 DOCKER_TAG=local
 DOCKER_TAG_DEV=local-dev
-DOCKER_WORKDIR=/opt/stactools-ephemeral
-DOCKER_NAMESPACE_PACKAGE_DIR=stactools/ephemeral
+DOCKER_WORKDIR=/opt/stactools-nrcan-radarsat1
+DOCKER_NAMESPACE_PACKAGE_DIR=stactools/nrcan_radarsat1
