+++ environment.yml
@@ -1,4 +1,4 @@
-name: stactools-ephemeral
+name: stactools-nrcan-radarsat1
 channels:
   - conda-forge
   - defaults
