+++ setup.cfg
@@ -1,15 +1,15 @@
 [metadata]
-name = stactools-ephemeral
-version = attr: stactools.ephemeral.__version__
+name = stactools-nrcan-radarsat1
+version = attr: stactools.nrcan_radarsat1.__version__
 description = PROVIDE DESCRIPTION HERE
 long_description = file: README.md
 long_description_content_type = text/markdown
 author = stac-utils
 author_email = stac@radiant.earth
-url = https://github.com/stactools-packages/ephemeral
+url = https://github.com/stactools-packages/nrcan-radarsat1
 project_urls =
-    Documentation = https://stactools-ephemeral.readthedocs.io/en/latest/
-    Issues = https://github.com/stactools-packages/ephemeral/issues
+    Documentation = https://stactools-nrcan-radarsat1.readthedocs.io/en/latest/
+    Issues = https://github.com/stactools-packages/nrcan-radarsat1/issues
 keywords =
     stactools
     pystac
