+++ tests/test_commands.py
@@ -2,13 +2,13 @@ import os.path
 from tempfile import TemporaryDirectory
 
 import pystac
-from stactools.ephemeral.commands import create_ephemeralcmd_command
+from stactools.nrcan_radarsat1.commands import create_nrcanradarsat1_command
 from stactools.testing import CliTestCase
 
 
 class CommandsTest(CliTestCase):
     def create_subcommand_functions(self):
-        return [create_ephemeralcmd_command]
+        return [create_nrcanradarsat1_command]
 
     def test_create_collection(self):
         with TemporaryDirectory() as tmp_dir:
@@ -18,7 +18,7 @@ class CommandsTest(CliTestCase):
             destination = os.path.join(tmp_dir, "collection.json")
 
             result = self.run_command(
-                ["ephemeralcmd", "create-collection", destination])
+                ["nrcanradarsat1", "create-collection", destination])
 
             self.assertEqual(result.exit_code,
                              0,
@@ -40,7 +40,7 @@ class CommandsTest(CliTestCase):
             # Example:
             destination = os.path.join(tmp_dir, "collection.json")
             result = self.run_command([
-                "ephemeralcmd",
+                "nrcanradarsat1",
                 "create-item",
                 "/path/to/asset.tif",
                 destination,
