+++ tests/test_module.py
@@ -1,8 +1,8 @@
 import unittest
 
-import stactools.ephemeral
+import stactools.nrcan_radarsat1
 
 
 class TestModule(unittest.TestCase):
     def test_version(self):
-        self.assertIsNotNone(stactools.ephemeral.__version__)
+        self.assertIsNotNone(stactools.nrcan_radarsat1.__version__)
